refactor(audio): move AudioDiagnosticsUI to its own folder
