+++ src/dotnet/UI.Blazor.App/Services/AudioDiagnostics/AudioContextDiagnostics.cs
@@ -1,0 +1,14 @@
+namespace ActualChat.UI.Blazor.App.Services;
+
+public sealed record AudioContextDiagnostics(
+    string Purpose,
+    string State,
+    double? SampleRate,
+    double? BaseLatencyMs,
+    double? OutputLatencyMs,
+    bool IsRunning,
+    bool IsMaintained,
+    bool IsUsed,
+    int RefCount,
+    bool IsReady,
+    string? BackgroundActivity);
