+++ src/dotnet/UI.Blazor.App/Services/AudioDiagnostics/AudioDiagnosticsUI.cs
@@ -1,7 +1,6 @@
 using ActualChat.Hosting;
 using ActualChat.UI.Blazor.App.Module;
 using ActualChat.UI.Blazor.Services;
-using Microsoft.JSInterop;
 
 namespace ActualChat.UI.Blazor.App.Services;
 
@@ -9,7 +8,6 @@ namespace ActualChat.UI.Blazor.App.Services;
 /// Gathers audio diagnostics for the Audio Diagnostics UI: the native audio-focus /
 /// session snapshot, and — on the web build only — the Web Audio playback state.
 /// </summary>
-// TODO: own folder
 // TODO: fix warnings
 // TODO: UIWorkerBase
 public sealed class AudioDiagnosticsUI(AppUIHub hub)
@@ -40,34 +38,3 @@ public sealed class AudioDiagnosticsUI(AppUIHub hub)
     public ValueTask ResumeContext()
         => JS.InvokeVoidAsync(JSResumeContextMethod);
 }
-
-/// <summary>
-/// Typed mirror of the TS <c>collectAudioPlaybackDiagnostics()</c> result.
-/// <see cref="Context"/> is null on native apps, which have no Web Audio pipeline.
-/// </summary>
-public sealed record AudioPlaybackDiagnostics(
-    AudioContextDiagnostics? Context,
-    AudioPlayerDiagnostics[] Players);
-
-public sealed record AudioContextDiagnostics(
-    string Purpose,
-    string State,
-    double? SampleRate,
-    double? BaseLatencyMs,
-    double? OutputLatencyMs,
-    bool IsRunning,
-    bool IsMaintained,
-    bool IsUsed,
-    int RefCount,
-    bool IsReady,
-    string? BackgroundActivity);
-
-public sealed record AudioPlayerDiagnostics(
-    string InternalId,
-    string? AuthorId,
-    string PlaybackState,
-    string BufferState,
-    double? PresentationLagMs,
-    double TargetBufferSizeMs,
-    double PlayingAt,
-    double BufferedDuration);
