+++ src/dotnet/UI.Blazor.App/Services/AudioDiagnostics/AudioPlayerDiagnostics.cs
@@ -1,0 +1,11 @@
+namespace ActualChat.UI.Blazor.App.Services;
+
+public sealed record AudioPlayerDiagnostics(
+    string InternalId,
+    string? AuthorId,
+    string PlaybackState,
+    string BufferState,
+    double? PresentationLagMs,
+    double TargetBufferSizeMs,
+    double PlayingAt,
+    double BufferedDuration);
